Rearranged the entire codebase, to fit a new architecture. this is the new Heirarchy Handlers- recieves input and events from other parts of the project and then calls the functions of the appropriate systems in response.
Systems-
makes use of sets and factorys, collections, and other systems to accomplish a given functionality.

Factories-
Provides the logic to create a given type of entity given some parameters.

Collections-
keeps a reference to all entities of a specific type and provides the methods to adding, removing and retrieving them.

Entities-
a game object which is the parent a list of child game objects, each child game object has exactly one purpose.